rename package .ui.view to .ui.viewholder
